precio type

+++ src/main/java/eu/rafaelaznar/bean/specificimplementation/ProductoSpecificBeanImplementation.java
@@ -131,7 +131,7 @@ public class ProductoSpecificBeanImplementation extends TableGenericBeanImplemen
         this.setCodigo(oResultSet.getString("codigo"));
         this.setDescripcion(oResultSet.getString("descripcion"));
         this.setExistencias(oResultSet.getInt("existencias"));
-        this.setPrecio(oResultSet.getInt("precio"));
+        this.setPrecio(oResultSet.getDouble("precio"));
         return this;
     }
